refactor: Move http_client out of the package holding only it
services/helper/ was left with one module after the helper.py split. A
directory is not a subject.
